@@ -3,10 +3,11 @@
   "version": "1.0.0",
   "description": "Discord bot with a collection of Phasmophobia utilities",
   "main": "dist/bot.js",
+  "type": "module",
   "scripts": {
-    "start": "cd dist && node bot.js",
+    "start": "cd dist && node --experimental-specifier-resolution=node bot.js",
     "build": "cd src && tsc",
-    "update_commands": "cd dist && railway run node updateSlashCommands.js"
+    "update_commands": "cd dist && node --experimental-specifier-resolution=node updateSlashCommands.js"
   },
   "repository": {
     "type": "git",
@@ -19,47 +20,38 @@
   },
   "homepage": "https://github.com/GuidoFe/phasmo-sidekick#readme",
   "dependencies": {
-    "@discordjs/builders": "^0.6.0",
-    "@discordjs/rest": "^0.1.0-canary.0",
+    "@discordjs/builders": "^1.2.0",
+    "@discordjs/rest": "^1.2.0",
     "@types/axios": "^0.14.0",
     "ansi-regex": ">=6.0.1",
-    "axios": "^0.22.0",
-    "color": "^4.0.1",
-    "dbots": "^9.0.1",
-    "discord-api-types": "^0.23.1",
-    "discord.js": "^13.1.0",
-    "dotenv": "^10.0.0",
-    "express": "^4.17.1",
-    "express-rate-limit": "^5.3.0",
-    "follow-redirects": ">=1.14.7",
-    "module-alias": "^2.2.2",
-    "nanoid": ">=3.1.31",
-    "node-fetch": "^2.6.6",
+    "axios": "^0.27.2",
+    "color": "^4.2.3",
+    "commonjs": "^0.0.1",
+    "dbots": "^11.0.0",
+    "discord-api-types": "^0.37.11",
+    "discord.js": "^14.5.0",
+    "dotenv": "^16.0.2",
+    "esm": "^3.2.25",
+    "express": "^4.18.1",
+    "express-rate-limit": "^6.6.0",
+    "follow-redirects": ">=1.15.2",
+    "nanoid": ">=4.0.0",
+    "node-fetch": "^3.2.10",
     "rest": "^2.0.0",
-    "topgg-autoposter": "^2.0.0"
+    "topgg-autoposter": "^2.0.1"
   },
   "devDependencies": {
-    "@babel/eslint-parser": "^7.15.7",
-    "@types/color": "^3.0.2",
+    "@babel/eslint-parser": "^7.19.1",
+    "@types/color": "^3.0.3",
     "@types/express-rate-limit": "^5.1.3",
-    "@types/node-fetch": "^2.5.12",
-    "@typescript-eslint/eslint-plugin": "^4.32.0",
-    "@typescript-eslint/parser": "^4.32.0",
-    "eslint": "^7.32.0",
+    "@types/node-fetch": "^2.6.2",
+    "@typescript-eslint/eslint-plugin": "^5.38.1",
+    "@typescript-eslint/parser": "^5.38.1",
+    "eslint": "^8.24.0",
     "eslint-config-google": "^0.14.0",
-    "follow-redirects": ">=1.14.7",
-    "mocha": "^9.1.1",
-    "nanoid": ">=3.1.31",
-    "typescript": "^4.4.3"
-  },
-  "_moduleAliases": {
-    "@modules": "dist/modules",
-    "@commands": "dist/commands",
-    "@adminCommands": "dist/admin_commands",
-    "@challenges": "challenges",
-    "@utils": "dist/utils.js",
-    "@constants": "dist/constants",
-    "@index": "dist/bot.js",
-    "@ghosts": "dist/ghosts"
+    "follow-redirects": ">=1.15.2",
+    "mocha": "^10.0.0",
+    "nanoid": ">=4.0.0",
+    "typescript": "^4.8.4"
   }
 }
