@@ -1,9 +1,9 @@
 import {SlashCommandBuilder} from '@discordjs/builders';
-import {SlashCommand, DataManager, Ghost} from '@modules';
-import utils = require('@utils');
-import {CommandInteraction, MessageEmbed, MessageAttachment} from 'discord.js';
-import {ghosts} from '@ghosts';
-import fs = require('fs');
+import {SlashCommand, DataManager, Ghost} from '../modules';
+import * as utils from '../utils';
+import {ChatInputCommandInteraction, EmbedBuilder} from 'discord.js';
+import {ghosts} from '../ghosts';
+import {readdirSync} from 'fs';
 
 export class GhostCommand extends SlashCommand {
     name = 'ghost';
@@ -15,22 +15,22 @@ export class GhostCommand extends SlashCommand {
         this.command = new SlashCommandBuilder()
             .setName(this.name)
             .setDescription(this.shortDescription)
-        let ghostOptions: [string, string][] = []
+        let ghostOptions: {name: string; value: string}[] = []
         ghosts.forEach((ghost, key) => {
-            ghostOptions.push([ghost.name, key])
+            ghostOptions.push({name: ghost.name, value: key})
         })
         this.command.addStringOption(option => 
                  option
                      .setName('type')
                      .setDescription("Ghost type")
                      .setRequired(true)
-                     .addChoices(ghostOptions)
+                     .addChoices(...ghostOptions)
         )
-        this.nThumbnails = fs.readdirSync('../res/images/ghosts').length
+        this.nThumbnails = readdirSync('../res/images/ghosts').length
     };
-    execute = async (interaction: CommandInteraction) => {
+    execute = async (interaction: ChatInputCommandInteraction) => {
         const ghost = ghosts.get(interaction.options.getString("type", true))!
-        const embed = new MessageEmbed()
+        const embed = new EmbedBuilder()
             .setTitle(ghost.name)
         let clues: {name: string, value: string, inline: boolean}[] = []
         ghost.clues.forEach((value, index) => clues.push({name: index == 0 ? "Evidence" : "\u200b", value: `${this.dataManager.constants.clueEmoji[value]} ${this.dataManager.constants.clueNames[value]}`, inline: true}))
@@ -38,12 +38,13 @@ export class GhostCommand extends SlashCommand {
                          {name: "Description", value: ghost.description, inline: false},
                          {name: "Strength", value: ghost.strength, inline: false},
                          {name: "Weakness", value: ghost.weakness, inline: false},
-                         {name: "Origin", value: ghost.flag ? `${ghost.flag} ${ghost.origin}` : ghost.origin, inline: false}
+                         {name: "Origin", value: ghost.flag ? `${ghost.flag} ${ghost.origin}` : ghost.origin, inline: false},
+			 {name: "Wiki", value: `[${ghost.name} page](${ghost.url})`, inline: false}
         ])
         embed.setColor(utils.randomVibrantColor())
         embed.setThumbnail('attachment://ghostie.png')
         //embed.setDescription(`**Origin**\n\n${ghost.flag ? `${ghost.flag} ${ghost.origin}` : ghost.origin}`)
-        embed.setFooter("From Phasmophobia Wiki, Wikipedia and Oxford Languages.")
+        embed.setFooter({text: "From Phasmophobia Wiki, Wikipedia and Oxford Languages."})
 //        const thumbnailUrl = `https://${process.env.RAILWAY_STATIC_URL}/res/images/ghosts/ghost${utils.getRandomInt(this.nThumbnails)}.png`
  //       embed.setThumbnail(thumbnailUrl)
 
