@@ -1,4 +1,4 @@
-import { Constants } from "@modules"
+import { Constants } from "./modules"
 export const constants: Constants = {
     items: [
         "📻 Spirit box",
@@ -26,7 +26,7 @@ export const constants: Constants = {
 
     inviteLink: "https://discord.com/api/oauth2/authorize?client_id=887086717587320852&scope=applications.commands",
     supportInviteLink: "https://discord.gg/4eJsE2byEY",
-    maps: [ "Tanglewood", "Edgefield", "Willow", "Ridgeview", "Grafton", "Bleasdale", "Highschool", "Prison", "Asylum", "Campfire" ],
+    maps: [ "Tanglewood", "Edgefield", "Willow", "Ridgeview", "Grafton", "Bleasdale", "Highschool", "Prison", "Sunny Meadows", "Sunny Meadows Restricted", "Maple Lodge", "Camp Woodwind"],
     clueNames: [
         "EMF 5", //0
         "Ghost Writing", //1
